pyproject.toml
[tool.poetry]
name = "django-client-framework"
version = "1.5.1"
description = ""
authors = ["Yu Li <ylilarry@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.4"
Django = "^4.0"
ipromise = "^1.9"
orjson = "^3.6.4"
deprecation = "^2.1.0"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
locust = "^2.2.1"
django-silk = "^4.1.0"
click = "^8.0.1"
coverage = "^5.5"
black = "^21.8b0"
djangorestframework-types = "^0.6.0"
django-types = "^0.8.1"
flake8 = "^3.9.2"
pylint = "^2.10.2"
autoflake = "^1.4"
isort = "^5.9.3"
schema = "^0.7.4"
django-stubs = "^1.9.0"
djangorestframework-stubs = "^1.4.0"
debugpy = "^1.5.1"
ipython = "^8.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dcf_clientschema = "django_client_framework.client_schema.cli:main"

[tool.pyright]
include = ["**/*.py"]
stubPath = './stubs'

[tool.isort]
profile = "black"
